refactor: adjust binding file structure
